% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimapR-helper.R
\name{clone_repo}
\alias{clone_repo}
\title{Clone a Git repository}
\usage{
clone_repo(repo_url, dest_dir)
}
\arguments{
\item{repo_url}{The URL of the Git repository.}

\item{dest_dir}{The destination directory where the repository will be cloned.}
}
\description{
This function clones a Git repository to a local directory.
}
\details{
This function requires Git to be installed on the system.
}
\note{
Ensure that Git is installed and available in the system PATH.
}
\examples{
\dontrun{
clone_repo("https://github.com/user/repo.git", "path/to/dest_dir")
}
}
