% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimapR-main.R
\name{minimap2}
\alias{minimap2}
\title{minimap2}
\usage{
minimap2(
  reference,
  query_sequences,
  output_file_prefix,
  a = TRUE,
  preset_string = "map-hifi",
  threads = 1,
  return_out = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{reference}{Reference genome to align the query sequences}

\item{query_sequences}{Query sequences to align to the reference genome}

\item{a}{Logical value to use the preset string with the -a flag}

\item{preset_string}{Preset string to use with the -x flag}

\item{threads}{Number of threads to use}

\item{verbose}{Logical value to print progress of the installation}

\item{...}{Additional arguments to pass to minimap2}

\item{output_file}{Output file to save the alignment results}
}
\value{
This function returns the line needed to add minimap2 to PATH
}
\description{
This function is a wrapper for the command line tool minimap2.
minimap2 is a long read sequencing alignment tool that is
used to align long reads to a reference genome.
}
