% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimapR-helper.R
\name{minimap2_install}
\alias{minimap2_install}
\title{minimap2_install}
\usage{
minimap2_install(source_directory, verbose = TRUE, return = FALSE)
}
\arguments{
\item{source_directory}{Source directory to install minimap2. Do not include minimap2 name in the
<<<<<<< HEAD
source directory. Note that this must be entered as a full path location.}
=======
source directoryro}
>>>>>>> main/pulipakabranch

\item{verbose}{Logical value to print progress of the installation}

\item{return}{This logical value causes the \code{minimap2_install} function to return the path of minimap2}
}
\value{
<<<<<<< HEAD
This function returns a character.
=======
Character value that is the path of the installed 'minimap2' tool.
>>>>>>> main/pulipakabranch
}
\description{
Install \code{minimap2} from Heng Li's github repository. If using a Windows operating system, installation of the MSYS2 Linux emulator is required.
}
\examples{
\dontrun{
install_dir <- file.path("/dir/to/install")
minimap2_path <- minimap2_install(source_directory = install_dir, verbose = FALSE)
}
}
