% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimapR-helper.R
\name{minimap2_installation}
\alias{minimap2_installation}
\title{minimap2_installation}
\usage{
minimap2_installation(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical value to print progress of the installation}
}
\value{
This function returns the path of the installed 'minimap2' tool (character).
}
\description{
This function prints installation instructions specific to the user's operating system.
}
\examples{
\dontrun{
minimap2_installation(verbose = FALSE)
}
}
