% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimapR-package.R
\docType{package}
\name{minimapR-package}
\alias{minimapR}
\alias{minimapR-package}
\title{minimapR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jake Reed \email{hreed.ga@gmail.com} (\href{https://orcid.org/0000-0002-8384-3593}{ORCID})

}
\keyword{internal}
