% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimapR-helper.R
\name{mm2_install}
\alias{mm2_install}
\title{minimap2_install}
\usage{
mm2_install(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical value to print progress of the installation. Default is \code{TRUE}. If set to \code{FALSE}, it will suppress the output messages during installation.}
}
\value{
If '\code{minimap2}' is not installed, this function installs it on windows, linux or macOS.
}
\description{
Install \code{minimap2} using conda or mamba. Conda or mamba is required for installation.
}
\examples{
\dontrun{
mm2_install(verbose = TRUE)
}
}
