% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimapR-helper.R
\name{samtools_check}
\alias{samtools_check}
\title{samtools_check}
\usage{
samtools_check(return = TRUE)
}
\arguments{
\item{return}{Logical value to return the path of samtools}
}
\value{
If '\code{samtools}' is installed, this function returns the path of samtools (character).
}
\description{
Check if samtools is installed
}
\examples{
samtools_check(return = TRUE)

}
