% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimapR-helper.R
\name{samtools_install}
\alias{samtools_install}
\title{samtools_install}
\usage{
samtools_install(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical value to print progress of the installation}
}
\description{
Install samtools with conda
}
\examples{
samtools_install()

}
