% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimapR-helper.R
\name{minimap2_check}
\alias{minimap2_check}
\title{minimap2_check}
\usage{
minimap2_check(return = TRUE)
}
\arguments{
\item{return}{Logical value to return the path of minimap2}
}
\value{
This function returns the path of minimap2 if installed
}
\description{
Check if minimap2 is installed
}
\examples{
minimap2_check(return == TRUE)

}
