% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimapR-helper.R
\name{minimap2_install}
\alias{minimap2_install}
\title{minimap2_install}
\usage{
minimap2_install(source_directory, verbose = TRUE)
}
\arguments{
\item{source_directory}{Source directory to install minimap2. Do not include minimap2 name in the
source directory}

\item{verbose}{Logical value to print progress of the installation}
}
\value{
This function returns the line needed to add minimap2 to PATH
}
\description{
Install minimap2 from Heng Li's github repository
}
\examples{
install_dir <- file.path("/dir/to/install")
path_line <- minimap2_install(source_directory = install_dir, verbose = FALSE)
}
