% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimapR-helper.R
\name{samtools_install}
\alias{samtools_install}
\title{samtools_install}
\usage{
samtools_install(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical value to print progress of the installation}
}
\value{
If '\code{samtools}' is not installed, this function installs it returns the path of the installed \code{'samtools'} tool (character).
}
\description{
Install samtools with conda
}
\examples{
\dontrun{
samtools_install()
}

}
